+++ CHANGELOG.md
@@ -71,6 +71,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - (GH #547) Fixed upload button being enabled when there are no files to upload.
 - (GH #549) Fixed changing project keeps loading previous project data.
 - (GH #550) Fixed changing project shows container from previous project.
+- Correctly set the global font to Museo sans
 
 ## [v2.0.1]